fix(react-mcp): refresh direct server resources
